+++ kmath-stat/src/jvmTest/kotlin/kscience/kmath/stat/SamplerTest.kt
@@ -1,4 +1,4 @@
-package kscience.kmath.prob
+package kscience.kmath.stat
 
 import kotlinx.coroutines.runBlocking
 import kotlin.test.Test